pyproject.toml
[tool.poetry]
authors = ["Daniel Schmidt <danischm@cisco.com>"]
description = "A CLI tool to perform syntactic and semantic validation of YAML files."
documentation = "https://github.com/netascode/iac-validate"
homepage = "https://github.com/netascode/iac-validate"
license = "LICENSE"
maintainers = ["Daniel Schmidt <danischm@cisco.com>"]
name = "iac-validate"
readme = "README.md"
repository = "https://github.com/netascode/iac-validate"
version = "0.2.5"

[tool.poetry.scripts]
iac-validate = "iac_validate.cli.main:main"

[tool.poetry.dependencies]
click = "^8.0.4"
errorhandler = "^2.0.1"
importlib-metadata = "^2.0.0"
python = "^3.8"
ruamel-yaml = ">0.16.10"
yamale = "^4.0.3"
mergedeep = "^1.3.4"

[tool.poetry.group.dev.dependencies]
ansible-core = "^2.13.2"
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.990"
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
toml = "^0.10.2"
types-pyyaml = "^6.0.12.2"

[tool.coverage.run]
source = ["iac_validate"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "@abstract", "except ImportError:"]
omit = ["*/__main__.py"]

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
# splits long import on multiple lines indented by 4 spaces
force_grid_wrap = 0
include_trailing_comma = true
indent = "    "
line_length = 88
multi_line_output = 3
use_parentheses = true
# will group `import x` and `from x import` of the same module.
combine_as_imports = true
default_section = "THIRDPARTY"
force_sort_within_sections = true
forced_separate = "tests"
known_first_party = ["iac_validate", "tests"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
markers = ["unit", "integration"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry_core>=1.0.0"]
